Better Git options, updates

+++ ProjectManager.js
@@ -1,17 +1,19 @@
 'use strict';
 
-const fs = require('fs-extra')
-	, url = require('url')
-	, Store = require('electron-store')
-	, moment = require('moment')
-	/* Git APIs */
-	, Blih = require('blih-api')
-	, GitHub = require('github')
+const fs = require('fs-extra');
+const path = require('path');
+const Store = require('electron-store');
+const moment = require('moment');
+const git = require('simple-git')();
+/* Git APIs */
+const Blih = require('blih-api');
+const GitHub = require('github');
 
 const data = new Store({ name: 'data' });
 const config = new Store({ name: 'config' }).store;
 
 class ProjectManager {
+
 	constructor (project) {
 		Object.assign(this, project);
 	}
@@ -31,7 +33,7 @@ class ProjectManager {
      		}
      	}
 		/* Check for name availability */
-		if (this.isProject() || this.isArchive()) {
+		if (data.store.projects.find(p => p.name === this.name)) {
 			throw new Error('This name is already taken by another project');
 		}
 		/* Check for directory availability */
@@ -42,51 +44,73 @@ class ProjectManager {
 		if (this.deadline && moment(this.deadline).isBefore(this.start)) {
 			throw new Error('The deadline cannot be after the start of the project.');
 		}
-
-		if (this.git.enabled) {
-			if (this.git.api) {
-				if (!this.git.username || !this.git.password) {
-					throw new Error('Git is enabled but credentials are missing');
-				}
-			} else {
-				throw new Error('Git is enabled but no API was specified');
-			}
-		}
-	}
-
-	/* TODO - review: if add remote fails, repo & dir are still here */
-	create (callback) {
-		if (this.git.enabled) {
-			this.createRepository((err) => {
-				if (!err) {
-					this.createDirectory((err) => {
-						if (!err) {
-							require('simple-git')(this.directory)
-								.init()
-								.addRemote('origin', this.git.remote, (err) => {
-									if (!err) {
-										this.save();
-									}
-									callback(err);
-								});
-						} else {
-							callback(err);
-						}
-					});
+		/* Check for git info consistency */
+		if (this.git.repo !== 'none') {
+			if (this.git.repo === 'new') {
+				if (this.git.api) {
+					// Info that will be sent to APIs.
+					if (this.git.api === 'blih' && !this.git.legacyUsername)
+						throw new Error('The "legacyUsername" field is mandatory to use the BLIH API.');
+					if (!this.git.username || !this.git.password)
+						throw new Error('You must provide credentials to create the repository.');
+					if (this.git.api !== 'blih')
+						delete this.git.legacyUsername;
 				} else {
-					callback(err);
+					throw new Error('You must specify an API to use to create the repository.');
 				}
-			});
+			} else if (this.git.repo === 'use') {
+				if (!this.git.remote)
+					throw new Error('No remote specified.');
+			} else {
+				throw new Error('Internal error.')
+			}
 		} else {
-			this.createDirectory((err) => {
-				if (!err) {
-					this.save();
-				}
-				callback(err);
-			});
+			delete this.git.api;
+			delete this.git.username;
+			delete this.git.legacyUsername;
+			delete this.git.password;
 		}
 	}
 
+	/* TODO - review: if add remote fails, repo & dir are still here . Use promises */
+	create (callback) {
+		this.createDirectory((err) => {
+			if (err) {
+				callback(err);
+			} else {
+				if (this.git.repo !== 'none') {
+					if (this.git.repo === 'new') {
+						this.createRepository((err) => {
+							if (!err) {
+								this.addRemote((err) => {
+									this.save();
+									callback();
+								});
+							} else {
+								callback(err);
+							}
+						});
+					} else {
+						this.addRemote((err) => {
+							this.pullRemote((err) => {
+								//console.log(err);
+								this.save();
+								callback();
+							});
+						});
+					}
+				} else {
+					this.save();
+					callback();
+				}
+			}
+		})
+		
+	}
+
+	/*
+	 * Create remote repository using git.api and sets git.remote accordingly.
+	 */
 	createRepository (callback) {
 		let api;
 		switch (this.git.api) {
@@ -106,7 +130,7 @@ class ProjectManager {
 					callback();
 				}).catch((err) => {
 					callback(err);
-				})
+				});
 				break ;
 			case 'blih':
 				api = new Blih(this.git.username, this.git.password)
@@ -114,7 +138,7 @@ class ProjectManager {
 				api.createRepository(this.name, (err, data) => {
 					console.log(err, data)
 					if (!err) {
-						this.git.remote = 'git@git.epitech.eu:/' + config.git.apis.blih.legacyUsername + '/' + this.name;
+						this.git.remote = 'git@git.epitech.eu:/' + this.git.legacyUsername + '/' + this.name;
 						callback();
 					} else {
 						callback(new Error(err));
@@ -140,34 +164,63 @@ class ProjectManager {
 		}
 	}
 
+	/*
+	 * Add git.remote to the project
+	 */
+	addRemote (callback) {
+		git.cwd(this.directory)
+			.init()
+			.addRemote('origin', this.git.remote, callback);
+	}
+
+	/*
+	 * Runs git pull origin master
+	 */
+	pullRemote (callback) {
+		git.cwd(this.directory)
+			.pull('origin', 'master', callback);
+	}
+
 	/***********/
 	/* Actions */
 	/***********/
 
-	archive (callback) {
-		let projects = data.store.projects.filter(p => p.name != this.name);
-		let archives = data.store.archives.filter(p => p.name != this.name);
+	move (callback, dest) {
+		let newdir = path.join(dest, this.name);
 
-		data.set('projects', projects);
-		archives.push(this);
-		data.set('archives', archives);
-		callback();
+		fs.move(this.directory, newdir)
+			.then(() => {
+				this.directory = newdir;
+				this.save();
+				callback();
+			}).catch((err) => {
+				callback(err)
+			});
+	}
+
+	archive (callback) {
+		this.active = false;
+		this.move(callback, config.archives.directory);
 	}
 
 	unarchive (callback) {
-		let projects = data.store.projects.filter(p => p.name != this.name);
-		let archives = data.store.archives.filter(p => p.name != this.name);
-
-		data.set('archives', archives);
-		projects.push(this);
-		data.set('projects', projects);
-		callback();
+		this.active = true;
+		this.move(callback, config.projects.directory);
 	}
 
 	save () {
 		let projects = data.get('projects');
 
-		projects.push(this);
+		if (!projects.find(p => p.name === this.name)) {
+			projects.push(this);
+		} else {
+			projects = projects.map(p => {
+				if (p.name === this.name)
+					Object.assign(p, this)
+				return p;
+			})
+		}
+
 		data.set('projects', projects);
 	}
 
@@ -189,51 +242,46 @@ class ProjectManager {
 
 	deleteGlobal (callback) {
 		let api;
-		switch (this.git.api) {
-			case 'github':
-				api = new GitHub();
 
-				api.authenticate({
-					type: config.git.apis.github.authentication,
-					username: this.git.username,
-					password: this.git.password
-				});
-				api.repos.delete({
-					owner: this.git.username,
-					repo: this.name
-				}).then((res) => {
-					this.deleteLocal(callback);
-				}).catch((err) => {
-					callback(err);
-				})
-				break ;
-			case 'blih':
-				api = new Blih(this.git.username, this.git.password);
+		if (!this.git.username || !this.git.password) {
+			callback(new Error('Username and password are required to authenticate.'))
+		} else {
+			switch (this.git.api) {
+				case 'github':
+					api = new GitHub();
 
-				api.deleteRepository(this.name, (err) => {
-					if (!err) {
+					api.authenticate({
+						type: config.git.apis.github.authentication,
+						username: this.git.username,
+						password: this.git.password
+					});
+					api.repos.delete({
+						owner: this.git.username,
+						repo: this.name
+					}).then((res) => {
 						this.deleteLocal(callback);
-					} else {
-						callback(new Error(err));
-					}
-				});
-				break ;
-			default:
-				callback(new Error('Cannot delete repository since this API is unknown'));
-				break ;
+					}).catch((err) => {
+						callback(err);
+					})
+					break ;
+				case 'blih':
+					api = new Blih(this.git.username, this.git.password);
+
+					api.deleteRepository(this.name, (err) => {
+						if (!err) {
+							this.deleteLocal(callback);
+						} else {
+							callback(new Error(err));
+						}
+					});
+					break ;
+				default:
+					callback(new Error('Cannot delete repository since this API is unknown'));
+					break ;
+			}
 		}
 	}
 
-	/***********/
-	/* Getters */
-	/***********/
-
-	isArchive () {
-		return data.store.archives.find(p => p.name === this.name) !== undefined;
-	}
-	isProject () {
-		return data.store.projects.find(p => p.name === this.name) !== undefined;
-	}
 }
 
 module.exports = ProjectManager;
